backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
asgiref
pyJWT
pytz
python-dotenv
PyMySQL